% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clisymbols-package.R
\docType{package}
\name{clisymbols-package}
\alias{clisymbols}
\alias{clisymbols-package}
\title{clisymbols: Unicode Symbols at the R Prompt}
\description{
A small subset of Unicode symbols, that are useful when building command line applications. They fall back to alternatives on terminals that do not support Unicode. Many symbols were taken from the 'figures' 'npm' package (see \url{https://github.com/sindresorhus/figures}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/clisymbols#readme}
  \item Report bugs at \url{https://github.com/r-lib/clisymbols/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Sindre Sorhus
}

}
\keyword{internal}
